requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
django-filter>=23.5
Pillow>=10.0
django-cors-headers>=4.3
dj-database-url>=2.1
